Clarify live image long press routing
