+++ mobile/apps/photos/lib/ui/viewer/file/detail_page.dart
@@ -422,7 +422,7 @@ class _BodyState extends State<_Body> {
           },
           backgroundDecoration: const BoxDecoration(color: Colors.black),
           qrDetectionsNotifier: _qrHelper?.qrDetectionsNotifier,
-          onLongPressStart:
+          onTextSelectionStart:
               flagService.ocrOverlayEnabled &&
                   widget.config.mode != DetailPageMode.minimalistic &&
                   !file.isLiveOrMotionPhoto
